Remove "colon specifier is experimental feature"
Same for "visibility is experimental feature"
With #7270 fixed, it's time to allow hackage uploads.

PWTExperimental is not used anymore, but for now I'm keeping it in case
we add other experimental features in the future, so we don't keep
removing and adding it (pushing CPP on users).

Since it isn't used though there is no way to run
cabal-testsuite/PackageTests/DuplicateExperimental
so I'm just removing that test.

Closes #6801