Drop outdated Hackage vs CPAN package counts
The package counts have grown to 15353 and 43351 respectively as of this
writing, hence the ratio has dropped from 5.6 to 2.8. Given the growth
of Hackage, the comparison does not seem as important anymore (does it
matter if you have 3 or 8 alternative libraries for the same purpose?).
Dropping package counts to avoid the need for continuous updates.