Clarify `runhaskell` in Cabal's guide
The original Cabal proposal https://www.haskell.org/cabal/proposal/ explained that `runhaskell`, in the proposal, was to refer to a symlink to `runhugs`, `runghc`, or `runnhc`. That explanation has been lost from the Cabal guide. Since the date of the original proposal, HUGS and NHC have fallen by the wayside. This pull request proposes two minor additions to restore the explanation:

1. A sentence "GHC provides the commands `runhaskell` and `runghc` (they are equivalent) to allow you to run Haskell programs without first having to compile them (scripts)." This wording closely follows that in the GHC user guide.

2. A clarification that `--help` provides information about the `Setup.hs` script (not the `runhaskell` command).