cabal init -i should autodetect author name and maintainer email (#8267)
* cabal init -i should autodetect author name and maintainer email (fix #8255)

Name and email are fetched from git config in the non-interactive init
already, and we plug that utility into the interactive one in this commit.

Testing becomes clunkier: the list of inputs have to hold an additional
string that will be used as the return value of
`Interactive.readProcessWithExitCode`, and here is why. Recall that this
method is used to call `git config` to perform autodetection. In the
real scenario (the `IO` instance of `Interactive`), it actually calls
out to `git`. In the testing scenario (the `PurePrompt` instance of
`Interactive`), the method simply returns the next element of the list
of pre-defined inputs. So, the list of inputs to every relevant test has
to be adjusted.

* Add test checking that username/email can be autodetected