Fix #8645: no assertion warning on special modes (help, version, etc.) (#8647)
* Fix #8654: no assertion warning on special modes (help, version, etc.)

This patch moves the `warnIfAssertionsAreEnabled` until after the main
command line argument parsing, and then only prints it on regular
operation, i.e. not when just the help or version of cabal was
demanded, or the command line could not be parsed.

Further, the warning is printed on stderr.  The previous attempt to
use `warn` does not work as expected, as the verbosity aperatshnik
isn't in place yet, so the warning cannot be suppressed by -v0.

Drive-by-shooting: straighten the convoluted control flow involving
`maybeScriptAndArgs`.  This IO computation was invoked "brutto"
before, when it is actually only needed in *one* branch of the case
distinction.

* PR #8647: changelog, fix text of warning

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>