Pass -this-unit-id when building executables
When there are multiple home units we need to be able to distinguish
between different executables. The way to do this is to pass the
`-this-unit-id` flag when compiling the executable.

Unfortunately this was broken before ghc-9.2 so there is a guard to
check that the compiler version is old enough. It's only critical we do
this from GHC-9.4 (when multiple home units was first supported) but
there's no harm in always passing this option.