Combine non-installable and non-upgradable package lists (fixes #8581)
This commit continues the work in ebe527185ea17497e782ae354b40090737c29ad2 to
combine the lists of non-installable and non-upgradable packages. When a package
is in the list, it means that the solver will only consider installed instances
of that package. This commit renames the list from "non-upgradable" to
"non-reinstallable" to clarify that the packages cannot be upgraded or
downgraded. There should be no change in behavior, except for the wording of
solver error messages.