Update RTD config for build.os config change (#9272)
* Update RTD config for build.os config change

See https://blog.readthedocs.com/use-build-os-config/ for details. We are in a temporary brownout with the old config key disabled, so RTD builds will be failing for the next few hours; but the change will become permanent next month.

* Remove old python version key

It's now in the build section.

* Downgrade Python back to 3.8

Otherwise need to update requirements.txt, which is beyond my pay grade.